removed duplicated logic and refactored to typescript
