+++ benchmarks/.gitignore
@@ -9,6 +9,7 @@ lerna-debug.log*
 .pnpm-debug.log*
 npm/
 node/node_benchmark.js
+utilities/*.js
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
@@ -274,5 +275,5 @@ _NCrunch*
 *.pdb
 dump.rdb
 
-# benchmark file with local configuration 
+# benchmark file with local configuration
 install_and_test_local_env.sh
